refator scripts as tools
